separate into component foler, util folder, separate handleSubmit into func
