+++ src/components/RegistrationForm.js
@@ -1,29 +1,22 @@
  // Render Prop
  import React from 'react';
  import { Formik, Form, Field, ErrorMessage } from 'formik';
+ import validation from '../utils/RegistrationFormValidation'
+
+ const handleSubmit = (values, { setSubmitting }) => {
+    setTimeout(() => {
+      alert(JSON.stringify(values, null, 2));
+      setSubmitting(false);
+    }, 400);
+  }
  
  const RegistrationForm = () => (
    <div>
-     <h1>Any place in your app!</h1>
+     <h1>Fake registration form!</h1>
      <Formik
        initialValues={{ email: '', password: '' }}
-       validate={values => {
-         const errors = {};
-         if (!values.email) {
-           errors.email = 'Required';
-         } else if (
-           !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i.test(values.email)
-         ) {
-           errors.email = 'Invalid email address';
-         }
-         return errors;
-       }}
-       onSubmit={(values, { setSubmitting }) => {
-         setTimeout(() => {
-           alert(JSON.stringify(values, null, 2));
-           setSubmitting(false);
-         }, 400);
-       }}
+       validate={validation}
+       onSubmit={handleSubmit}
      >
        {({ isSubmitting }) => (
          <Form>
